minor changes
- moved files around
- split functions in 2 files
