+++ frontend/utility/authFunctions.js
@@ -57,22 +57,3 @@ export async function registerUser(user) {
         throw err
     }
 }
-
-export async function getUserById(id) {
-
-    try {
-        // API CALL TO REGISTER USER
-        const response = await fetch(apiBase + '/'+id)
-
-        if (!response.ok) {
-            throw new Error("User does not exist");
-        }
-
-        const data = await response.json()
-
-        return data.user;
-    } catch(err) {
-        throw err
-    }
-
-}
